requirements.txt
PyYAML
ffmpy
BigQuery-Python
google-cloud-pubsub
boto3
gcloud
